% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_by_year_length.R
\name{adjust}
\alias{adjust}
\title{Adjust census data that uses the Chinese calendar}
\usage{
adjust(raw_data, year)
}
\arguments{
\item{raw_data}{vector of single year birth survivors}

\item{year}{year of census}
}
\value{
Returns a vector of birth survivor by single year, adjusted by year length in the Chinese calendar
}
\description{
Adjust census data that uses the Chinese calendar
}
\examples{
file = read_excel("C:/Users/alber/OneDrive/Desktop/Calendar Research/Census Data/Taiwan/Taiwan1905.xlsx", sheet=2)[1:101,]
data = as.data.frame(file)
adjust(data$Male,1905)
}
