% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leap_year_PAR.R
\name{leap_years_apar}
\alias{leap_years_apar}
\title{Average period age ratio of all Chinese leap years}
\usage{
leap_years_apar(x, year)
}
\arguments{
\item{x}{Vector of birth survivors}

\item{year}{year of census}
}
\value{
average period age ratio of all leap years
}
\description{
Average period age ratio of all Chinese leap years
}
\examples{
leap_years_apar(data$Female, 1905)
}
