% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAR_3.R
\name{par_3}
\alias{par_3}
\title{3-Year Period Age Ratio}
\usage{
par_3(x)
}
\arguments{
\item{x}{Vector of birth survivors, from age 0 to 100}
}
\value{
3 Year Period Age Ratio of ages 2 to 99. Note that the first and last element equals 0.
}
\description{
3-Year Period Age Ratio
}
