% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_par.R
\name{plot_par}
\alias{plot_par}
\title{Visualize period age ratio of Chinese leap years}
\usage{
plot_par(x, year, start = 1, end = 90, ...)
}
\arguments{
\item{x}{Vector of birth survivors, from age 0 to 100}

\item{year}{Year of Census}
}
\value{
a visualization of Chinese leap year period age ratio
}
\description{
Visualize period age ratio of Chinese leap years
}
\examples{
data(taiwan1905)
plot_par(taiwan1905$Total,1905)
}
