% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taiwan1915}
\alias{taiwan1915}
\title{1915 Taiwanese census data with only the local Taiwanese population (Japanese and Chinese excluded)}
\format{
A data frame with 101 rows and 4 variables:
\describe{
  \item{Age}{Age from 0 to 100}
  \item{Total}{Total number of birth survivors with a given age}
  \item{Male}{Total number of male birth survivors with a given age}
  \item{Female}{Total number of female birth survivors with a given age}
}
 @examples data(taiwan1915)
}
\usage{
taiwan1915
}
\description{
A dataset containing the number of birth survivors from each age group, broken down by gender.
}
\keyword{datasets}
