% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{year_length}
\alias{year_length}
\title{Length breakdown of each Chinese year in the western calendar}
\format{
A data frame with 197 rows and 5 variables:
\describe{
  \item{Chinese.Year}{The year according to the Chinese calendar. It starts roughly a month after the western calendar and has length varying between 353 and 385 days.}
  \item{Number of Days}{Length of the Chinese Year, varies from 353 to 385.}
  \item{Days in Current Western Year}{Number of days that Chinese year overlaps with the same western year}
  \item{Days in Next Western year}{Equals to Number of Days minus Days in Current Western Year}
  \item{Days from Previous Western year}{Equals to Days in Next Western year of the previous row}
}
 @examples data(year_length)
}
\usage{
year_length
}
\description{
A dataset containing the number of days that each Chinese year has that falls in each western year, going back to 1804.
}
\keyword{datasets}
