% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{years}
\alias{years}
\title{Collection of all Chinese leap years and their lengths}
\format{
A data frame with 197 rows and 5 variables:
\describe{
  \item{Year}{Th Chinese year}
  \item{Days}{Length of the Chinese Year, varies from 353 to 385.}
}
 @examples data(years)
}
\usage{
years
}
\description{
A dataset containing the length of each Chinese leap year from 1800 to 1982.
}
\keyword{datasets}
